Make settings utahId default to boolean
Having header's settings.utahId default to undefined
was causing some logic to misfire. It is now a required
field in the Settings type, but the default settings set it
to false.

With action items on, utah id set to undefined, and a main menu,
the utahId button/actionItem was not showing in mobile.