Make sure pageUrls match their menu paths
The added unit test REQUIRES they match. This may be too
harsh? But probably not. If they HAVE to match, then why not
reconstruct them programmatically? Menus and pages
organization may need a refactor so that they do always match.
But that is out of the scope of this commit.

UDS-895