Add TableContextConsumer component
This component allows generic use of the Table Context inline
with the table. React suggests not using SomeContext.consumer
so we wrapped our own useContext in a Consumer. The thought
is that a library should allow generic usage like this, though
an app should consider creating a custom component to get the context.

Co-authored-by: Joseph Sharp <jmsharp@utah.gov>

UDS-678