Remove esLintPlugin from vite local dev
The esLintPlugin was causing the local HMR vite to be
SUPER slow (10seconds vs 5 seconds). By removing
eslint, vite now is more usable. BUT it won't die if there
are eslint errors. We do want to have to fix eslint errors
so we are pushing the error checking to the build process.
We won't get live eslint errors, but it will still correctly bite us.

The eslint-build.yml file now runs on pull-requests to make sure
there are no eslint errors.