Add child menu item for main menu item
If the main menu item has children AND a link, then
add a child item for it like how the mobile menu does
this. It was initially thought only mobile needed it, but
ipads are mobile and don't trigger mobile view and need it.
WordPress header plugin needs this too.

UDS-1397