fix: Make main menu actionFunctionUrl items be buttons
If a main menu item is both a menu AND an actionFunctionUrl,
then the developer is up in the night. But main men
items that have a children menus must be a button so
that screen readers read them correctly. If they are hrefs
then the screen reader doesn't say it the same way as other menus with children.

UDS-1684